[Silver I] Title: 곱셈, Time: 0 ms, Memory: 2020 KB -BaekjoonHub

+++ 백준/Silver/1629. 곱셈/곱셈.cc
@@ -3,7 +3,7 @@ using namespace std;
 
 long long A, B, C;
 
-long long power(long long n)
+int power(int n)
 {
 	if (n == 1)
 		return (A % C);
